fix(frontend): copy button styling fixes
